move test graph to ns pkg
